refactor: remove chord recognition API and related files; update voice separation API and requirements
