--- chord_api.py
@@ -1,104 +1,0 @@
-import os
-import tempfile
-import numpy as np
-import librosa
-from dotenv import load_dotenv
-from fastapi import FastAPI, File, UploadFile
-from fastapi.responses import JSONResponse
-from fastapi.middleware.cors import CORSMiddleware
-
-from SongChordRecognizer_Pipeline.DataPreprocessor import DataPreprocessor
-from SongChordRecognizer_Training.Models import CRNN_basic_WithStandardScaler
-from SongChordRecognizer_Training.Spectrograms import cqt_spectrogram
-from SongChordRecognizer_Pipeline.KeyRecognizer import KeyRecognizer
-
-# Load environment variables
-load_dotenv()
-CRNN_MODEL_PATH = os.getenv("CRNN_MODEL_PATH")
-CRNN_SCALER_PATH = os.getenv("CRNN_SCALER_PATH")
-
-# Constants
-SAMPLE_RATE = int(os.getenv("SAMPLE_RATE", 22050))
-HOP_LENGTH = int(os.getenv("HOP_LENGTH", 512))
-N_FRAMES = int(os.getenv("N_FRAMES", 1000))
-
-# Initialize FastAPI app
-app = FastAPI()
-
-# Enable CORS
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["*"],
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
-
-
-@app.get("/health")
-async def health():
-    return JSONResponse(content={"status": "ok"})
-
-
-@app.post("/ai/predict-chord")
-async def predict_chord(file: UploadFile = File(...)):
-    """
-    Predict the main chord, full chord sequence, and key from an uploaded audio file.
-    """
-    tmp_path = None
-    try:
-        # Save the uploaded file temporarily
-        with tempfile.NamedTemporaryFile(delete=False, suffix=".wav") as tmp:
-            tmp.write(await file.read())
-            tmp_path = tmp.name
-
-        # Load audio
-        y, sr = librosa.load(tmp_path, sr=SAMPLE_RATE)
-
-        # Load model and scaler
-        model = CRNN_basic_WithStandardScaler()
-        model.load(CRNN_MODEL_PATH, CRNN_SCALER_PATH)
-
-        # Preprocess the audio for prediction
-        x = DataPreprocessor.sequence_preprocess(
-            waveform=y,
-            sample_rate=sr,
-            hop_length=HOP_LENGTH,
-            n_frames=N_FRAMES,
-            spectrogram_generator=cqt_spectrogram,
-            norm_to_C=False,
-        )
-
-        # Predict chords
-        predictions = model.predict(x)
-        chord_indices = predictions.argmax(axis=2).flatten()
-        # Main chord (most frequent)
-        main_chord_index = np.bincount(chord_indices).argmax()
-        main_chord = DataPreprocessor.chord_indices_to_notations([main_chord_index])[0]
-        chord_sequence = DataPreprocessor.chord_indices_to_notations(chord_indices)
-        # Key detection (music key, not just most frequent chord)
-        chords, counts = np.unique(chord_indices, return_counts=True)
-        chord_counts = dict(zip(chords, counts))
-        key = KeyRecognizer.estimate_key(chord_counts)
-
-        return JSONResponse(
-            content={
-                "key": key,
-                "main_chord": main_chord,
-                "chord_sequence": chord_sequence,
-            }
-        )
-
-    except Exception as e:
-        return JSONResponse(
-            content={"error": f"Internal error: {str(e)}"}, status_code=500
-        )
-    finally:
-        if tmp_path and os.path.exists(tmp_path):
-            os.remove(tmp_path)
-
-
-if __name__ == "__main__":
-    import uvicorn
-
-    uvicorn.run(app, host="0.0.0.0", port=8001)
