+++ requirements.txt
@@ -1,8 +1,16 @@
 tensorflow
 keras
+numpy
+scikit-learn
+matplotlib
+librosa
+mir_eval
+joblib
+sklearn-crfsuite
+fastapi
+uvicorn
+gunicorn
+python-dotenv
+python-multipart
 sounddevice
 scipy
-uvicorn
-librosa
-fastapi
-python-multipart
