--- requirements_chord.txt
@@ -1,15 +1,0 @@
-tensorflow==2.13.0
-keras==2.13.1
-numpy==1.23.5
-scikit-learn
-matplotlib==3.3.4
-librosa==0.8.1
-mir_eval==0.6
-fastapi==0.78.0
-uvicorn==0.17.6
-python-dotenv==0.19.2
-python-multipart==0.0.5
-joblib==1.0.1
-sklearn-crfsuite==0.3.6
-sounddevice
-scipy
