--- requirements_voice.txt
@@ -1,8 +1,0 @@
-fastapi==0.78.0
-uvicorn==0.17.6
-tensorflow
-keras
-sounddevice
-soundfile
-scipy 
-python-multipart
