+++ sample.env
@@ -1,5 +1,16 @@
-CRNN_MODEL_PATH=/Users/triductran/SpartanDev/my-work/datn/HarmonySeeker-ai/SongChordRecognizer_Pipeline/models/original_crnn.h5
+# Song Chord Recognizer parameter
+CRNN_MODEL_PATH=/Users/triductran/SpartanDev/my-work/datn/HarmonySeeker-ai/SongChordRecognizer_Pipeline/models/original_crnn.keras
 CRNN_SCALER_PATH=/Users/triductran/SpartanDev/my-work/datn/HarmonySeeker-ai/SongChordRecognizer_Pipeline/models/original_preprocessor.bin
+MLP_MODEL_PATH=/Users/triductran/SpartanDev/my-work/datn/HarmonySeeker-ai/SongChordRecognizer_Pipeline/my_models/mlp22_isophonics_206songs_noNormC.model
 SAMPLE_RATE=22050
 HOP_LENGTH=512
 N_FRAMES=1000
+
+# Voice Separator parameters
+VOICE_MODEL_PATH=/Users/triductran/SpartanDev/my-work/datn/HarmonySeeker-ai/VoiceSeparator_Pipeline/model/model_jax_20_epoch.keras
+VOICE_CHUNK_SIZE=65024
+VOICE_STFT_N_FFT=2048
+VOICE_STFT_HOP_LENGTH=512
+VOICE_N_SUBBANDS=4
+VOICE_N_INSTRUMENTS=1
+VOICE_TARGET_SAMPLE_RATE=16000
